+++ src/application/App.tsx
@@ -1,28 +1,21 @@
 import 'react-contexify/dist/ReactContexify.css';
 
 import { MenuProvider } from 'kmenu';
-import React, { useCallback, useLayoutEffect, useState } from 'react';
-import { MdKeyboard, MdSettings } from 'react-icons/md';
-import { ImperativePanelGroupHandle, ImperativePanelHandle, Panel, PanelGroup } from 'react-resizable-panels';
-import { useEventListener, useLocalStorage, useWindowSize } from 'usehooks-ts';
+import React, { useCallback, useLayoutEffect } from 'react';
+import { ImperativePanelGroupHandle, Panel, PanelGroup } from 'react-resizable-panels';
+import { useLocalStorage, useWindowSize } from 'usehooks-ts';
 
-import { VerticalResizeHandle } from '../components/VerticalResizeHandle';
 import { emitEvent } from '../events';
-import { ChatbotPanel } from '../features/ai/sidebar/ChatPanel';
-import { RewriterPanel } from '../features/ai/sidebar/RewriterPanel';
 import { ReferencesDropZone } from '../features/references/components/ReferencesDropZone';
-import { ReferencesPanel } from '../features/references/sidebar/ReferencesPanel';
 import { useDebouncedCallback } from '../hooks/useDebouncedCallback';
-import { noop } from '../lib/noop';
 import { SettingsModalOpener } from '../settings/SettingsModalOpener';
 import { ApplicationFrame } from '../wrappers/ApplicationFrame';
 import { ContextMenus } from '../wrappers/ContextMenus';
 import { EventsListener } from '../wrappers/EventsListener';
 import { CommandPalette } from './commands/CommandPalette';
 import { MainPanel } from './components/MainPanel';
-import { SideBar } from './components/SideBar';
-import { ExplorerPanel } from './sidebar/ExplorerPanel';
-import { BotIcon, FilesIcon, PenIcon, ReferencesIcon } from './sidebar/icons';
+import { LeftSidePanelWrapper } from './sidebar/LeftSidePanelWrapper';
+import { RightSidePanelWrapper } from './sidebar/RightSidePanelWrapper';
 
 export function App() {
   const panelRef = React.createRef<ImperativePanelGroupHandle>();
@@ -87,122 +80,5 @@ export function App() {
         <SettingsModalOpener />
       </ReferencesDropZone>
     </EventsListener>
-  );
-}
-
-type PrimarySideBarPane = 'Explorer' | 'References';
-function LeftSidePanelWrapper() {
-  const leftPanelRef = React.useRef<ImperativePanelHandle>(null);
-  const [primaryPaneCollapsed, setPrimaryPaneCollapsed] = useState(false);
-  const [primaryPane, setPrimaryPane] = useState<PrimarySideBarPane>('Explorer');
-
-  const handleSideBarClick = (selectedPane: PrimarySideBarPane) => {
-    if (selectedPane === primaryPane) {
-      // Toggle collapsing
-      setPrimaryPaneCollapsed(!primaryPaneCollapsed);
-    } else {
-      // Always open the sidebar
-      setPrimaryPaneCollapsed(false);
-    }
-    setPrimaryPane(selectedPane);
-  };
-
-  // Configure keyboard shortcuts to open/close side panel
-  useEventListener('keydown', (e) => {
-    if (e.metaKey) {
-      switch (e.key.toLowerCase()) {
-        case '1':
-          return handleSideBarClick('Explorer');
-        case '2':
-          return handleSideBarClick('References');
-      }
-    }
-  });
-
-  React.useEffect(() => {
-    if (primaryPaneCollapsed) {
-      leftPanelRef.current?.collapse();
-    } else {
-      leftPanelRef.current?.expand();
-    }
-  }, [leftPanelRef, primaryPaneCollapsed]);
-
-  const openSettings = React.useCallback(() => emitEvent('refstudio://menu/settings'), []);
-
-  return (
-    <>
-      <SideBar
-        activePane={primaryPaneCollapsed ? null : primaryPane}
-        footerItems={[
-          // TODO: Implement Keybinds screen
-          {
-            label: 'Keybinds',
-            Icon: <MdKeyboard aria-label="Keybindings" size="24" />,
-            onClick: noop,
-          },
-          {
-            label: 'Settings',
-            Icon: <MdSettings aria-label="Settings" size="24" />,
-            onClick: openSettings,
-          },
-        ]}
-        items={[
-          { pane: 'Explorer', Icon: <FilesIcon /> },
-          { pane: 'References', Icon: <ReferencesIcon /> },
-        ]}
-        onItemClick={handleSideBarClick}
-      />
-      {!primaryPaneCollapsed && <div className="h-full w-[1px]" style={{ backgroundColor: '#eff1f4' }} />}
-      <Panel collapsible order={1} ref={leftPanelRef} onCollapse={(collapsed) => setPrimaryPaneCollapsed(collapsed)}>
-        {primaryPane === 'Explorer' && <ExplorerPanel />}
-        {primaryPane === 'References' && <ReferencesPanel />}
-      </Panel>
-      <VerticalResizeHandle />
-    </>
-  );
-}
-
-type SecondarySideBarPane = 'Rewriter' | 'Chatbot';
-function RightSidePanelWrapper() {
-  const rightPanelRef = React.useRef<ImperativePanelHandle>(null);
-  const [secondaryPaneCollapsed, setSecondaryPaneCollapsed] = useState(true);
-  const [secondaryPane, setSecondaryPane] = useState<SecondarySideBarPane>('Rewriter');
-
-  const handleSideBarClick = (selectedPane: SecondarySideBarPane) => {
-    if (selectedPane === secondaryPane) {
-      // Toggle collapsing
-      setSecondaryPaneCollapsed(!secondaryPaneCollapsed);
-    } else {
-      // Always open the sidebar
-      setSecondaryPaneCollapsed(false);
-    }
-    setSecondaryPane(selectedPane);
-  };
-
-  React.useEffect(() => {
-    if (secondaryPaneCollapsed) {
-      rightPanelRef.current?.collapse();
-    } else {
-      rightPanelRef.current?.expand();
-    }
-  }, [rightPanelRef, secondaryPaneCollapsed]);
-
-  return (
-    <>
-      <VerticalResizeHandle />
-      <Panel collapsible order={3} ref={rightPanelRef} onCollapse={(collapsed) => setSecondaryPaneCollapsed(collapsed)}>
-        {secondaryPane === 'Rewriter' && <RewriterPanel />}
-        {secondaryPane === 'Chatbot' && <ChatbotPanel />}
-      </Panel>
-      {!secondaryPaneCollapsed && <div className="h-full w-[1px]" style={{ backgroundColor: '#eff1f4' }} />}
-      <SideBar
-        activePane={secondaryPaneCollapsed ? null : secondaryPane}
-        items={[
-          { pane: 'Rewriter', Icon: <PenIcon /> },
-          { pane: 'Chatbot', Icon: <BotIcon /> },
-        ]}
-        onItemClick={handleSideBarClick}
-      />
-    </>
   );
 }
